Revert "Kacper/dev 14"
