--- src/static/main/edit_recipes_styles.css
@@ -1,250 +1,0 @@
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-/* Firefox */
-input[type=number] {
-  -moz-appearance: textfield;
-}
-
-h3 {
-  font-weight:300;
-  margin-bottom: 5px;
-}
-
-h5 {
-    font-size: small;
-    font-weight: 400;
-    margin-bottom: 1px;
-}
-
-.containerContent{
-    position: relative;
-    z-index: 1;
-}
-
-#recipesDisplay{
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    row-gap: 2.5%;
-}
-
-#filters{
-    width: 25%;
-    overflow: auto;
-}
-
-
-.recipeDisplayed {
-    width: 26.5%;
-    max-height: 50%;
-    overflow: initial;
-    margin-left: 1%;
-    border: 1px gray solid;
-}
-
-.containerContent{
-    padding-bottom: 2.5%;
-}
-
-.in{
-    overflow: auto;
-}
-
-
-#left_bar{
-    position: relative;
-}
-
-#right_arrow_box{
-    position: absolute;
-    left: 2020%;
-    top: 96%;
-    z-index: 1;
-}
-
-#left_arrow_box{
-    position: absolute;
-    top:96%;
-    right: 50%;
-}
-
-.arrow_box{
-    width: 5%;
-    display: flex;
-    position: relative;
-}
-
-.arrow{
-    width: 25px;
-    height: 25px;
-    position: absolute;
-    top: 97%;
-    cursor: pointer;
-}
-
-#right_arrow{
-    position: relative;
-    right: 50%;
-}
-
-#left_arrow{
-    position: relative;
-    left: 50%;
-}
-
-
-.range_filter{
-    display: flex;
-}
-
-.separator{
-    margin-left: 2.5%;
-    margin-right: 2.5%;
-    text-align: center;
-}
-
-.unit{
-    position: absolute;
-    right: 11%;
-    top: 3%;
-}
-
-.range-input{
-    position: relative;
-}
-
-.filters{
-    width: 100%;
-    margin-bottom: 1%;
-    height: 100%;
-}
-
-#filters_id{
-    display: grid;
-}
-
-.input_box{
-    margin-bottom: 6%;
-}
-
-.number_filter{
-    padding-right: 25%;
-}
-
-@media (max-width:1080px){
-    [class*="number_filter"]{
-        padding-right: 35%;
-    }
-}
-
-.unit_cel{
-    right: 9%;
-}
-
-.filter_btn{
-    margin-bottom: 6%;
-}
-
-#createRecipeImg{
-    width: 60%;
-    height: 50%;
-    cursor: pointer;
-}
-
-#createRecipe{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#icons{
-    width: 5%;
-    margin-top: 2%;
-    position: relative;
-    max-height: 47%;
-}
-
-.ico_box{
-    width: 100%;
-    height: 12%;
-    background-color: white;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: absolute;
-    border-radius: 13%;
-}
-
-.ico{
-    width: 70%;
-    height: 70%;
-    cursor: pointer;
-}
-
-.pencil_ico{
-    top:20%;
-}
-
-.trash_ico{
-    top:40%;
-}
-
-.to_blur{
-  -webkit-filter: blur(2px);
-  -moz-filter: blur(2px);
-  -o-filter: blur(2px);
-  -ms-filter: blur(2px);
-  filter: blur(2px);
-}
-
-#addRecipePage{
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    display: none;
-    z-index: 0;
-    align-items: center;
-    justify-content: center;
-    filter: none;
-    overflow: hidden;
-}
-
-
-#content_to_blur{
-    display: flex;
-}
-
-.recipeDisplayedEdit{
-    height: 90%;
-    width: 38%;
-    border: 1px solid black;
-}
-
-#recipeEdit{
-    width: 100%;
-    height: 100%;
-}
-
-#icons_edit{
-    position: absolute;
-    left: 110%;
-    height: 90%;
-    width: 10%;
-}
-
-#create_recipe_id{
-    position: relative;
-}
-
-#recipeEdit{
-    overflow: visible;
-}
-
-.ico_edit{
-    width: 60%;
-    height: 80%;
-}
