--- src/static/main/home_styles.css
@@ -1,6 +1,0 @@
-
-
-.recipeDisplayed {
-    width: 33.33%; 
-    height: 100%;
-}
