+++ src/static/main/scripts/home_scripts.js
@@ -1,5 +1,5 @@
 var recipesContainer = document.getElementById("recipesDisplay");
-
+var machinInfoContainer = document.getElementById("infoBar");
 
 
 function loadRecipes(request_id, fetch_nexts, csrf_token) {
@@ -19,6 +19,26 @@ function loadRecipes(request_id, fetch_nexts, csrf_token) {
     requestData.send(JSON.stringify({ "from": request_id, "range":fetch_nexts}));
 
 }
+
+
+function loadMachineInfo(csrf_token) {
+    var requestData = new XMLHttpRequest();
+    requestData.responseType = "text";
+    requestData.addEventListener("load", function () {
+        if (requestData.status == 200) {
+            machinInfoContainer.innerHTML = requestData.responseText
+        }
+        else {
+            console.log('Request error')
+        }
+    }, {once : true});
+    requestData.open("post", window.location.href+'machineList');
+    requestData.setRequestHeader("X-CSRFToken", csrf_token);
+    requestData.send();
+
+}
+
+
 
 
 var toggled = undefined
