--- src/templates/main/edit_recipe.html
@@ -1,84 +1,0 @@
-{% load static %}
-
-<div class="centerIn recipeDisplayedEdit">
-    <div id="recipeEdit" class="in" style="width: 95%; height: 97%; border-radius: 5px; background-color: rgb(255,255,255); " onclick="toggle_select('recipesDisplayed{{recipe.id}}')">
-        <form method="POST" id="create_recipe_id" onsubmit="event.preventDefault();submit_recipe(this);">  {% csrf_token %}
-            <div class="recipeEditTitle centerIn" style="width: 100%; height: 19%;">
-                <div class="teaName in"> 
-                    <div>Nazwa</div>
-                    <div class="input_box_edit"> {{create_filters.recipe_name}} </div>
-                </div>
-            </div>
-            <div class="recipeEditBorder centerIn" style="width: 100%; height: 1%;">
-                <div class="in borderSec"></div>
-            </div>
-            <div class="recipeEditContent" style="width: 100%; height: 80%;">
-                <div class="recipeEditIngredients centerIn" style="width: 100%; height: auto;">
-                    <div class="in" style="width: 100%; height: 100%">
-                        <div class="dataTitle dataPos" style="width: 100%; height: 30%">Skłdaniki</div>
-                        <div class="data" style="width: 100%; height: 80%; display: flex;">
-                            <div class="names" style="display: flex; flex-wrap: wrap; width: 69%; height: 100%;">
-                                <div style="height: 20px;">Woda</div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.tea_name}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ingredient1}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ingredient2}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ingredient3}} </div>
-                                
-                            </div>
-                            <div style="height: auto; width: 1;"> <div style="height: 10%;"></div> <div class="borderSec-ver"></div></div>
-                            <div class="quantity" style="display: flex; flex-wrap: wrap; width: 30%; height: 100%;">
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.water}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.tea_quan}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ing1_ammount}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ing2_ammount}} </div>
-                                <div class="input_box_edit" style="height: 20px; width: 100%;"> {{create_filters.ing3_ammount}} </div>
-                                
-        
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="recipeEditTemperatures centerIn" style="width: 100%; height: auto;">
-                    <div class="in" style="width: 100%; height: 100%">
-                        <div class="dataTitle dataPos" style="width: 100%; height: 20%">Temperatury</div>
-                        <div class="data" style="width: 100%; height: 80%; display: flex;">
-                            <div class="names" style="width: 69%; height: 100%;">
-                                <div>Parzenie</div>
-        
-                            </div>
-                            <div style="height: auto; width: 1;"> <div style="height: 5%;"></div> <div class="borderSec-ver"></div></div>
-                            <div class="quantity" style="width: 30%; height: 100%;">
-                                <div class="input_box_edit"> {{create_filters.brewing_temperature}} </div>
-        
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="recipeEditTimes centerIn" style="width: 100%; height: auto;">
-                    <div class="in" style="width: 100%; height: 100%">
-                        <div class="dataTitle dataPos" style="width: 100%; height: 20%">Czasy</div>
-                        <div class="data" style="width: 100%; height: 80%; display: flex;">
-                            <div class="names" style="width: 69%; height: 100%;">
-                                <div>Parzenie</div>
-                                <div>Mieszanie</div>
-        
-                            </div>
-                            <div style="height: auto; width: 1;"> <div style="height: 5%;"></div> <div class="borderSec-ver"></div></div>
-                            <div class="quantity" style="width: 30%; height: 100%;">
-                                <div class="input_box_edit"> {{create_filters.brewing_time}} </div>
-                                <div class="input_box_edit"> {{create_filters.mixing_time}} </div>
-        
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div id="icons_edit">
-                    <div id="save_icon_edit" class="save_icon_edit ico_box_edit"> <input type="image" src="{% static 'main/media/heart.svg' %}" alt="" class="ico_edit" /></div>
-                    <div id="return_icon_edit"  class="return_icon_edit ico_box_edit " onclick="close_create_recipe()"> <image src="{% static 'main/media/trash.svg' %}" alt="" class="ico_edit"></div>
-                </div>
-            </div>
-
-        </form>
-    </div>
-</div>
-
