--- src/templates/main/edit_recipes.html
@@ -1,77 +1,0 @@
-{%extends "main/base.html"%}
-
-{% load static %}
-
-{% block additional_styles %}
-    <link rel="stylesheet" type="text/css" href="{% static 'main/edit_recipes_styles.css' %}">
-{% endblock additional_styles %}
-
-{%block main_content%}
-    <div id="addRecipePage">
-    {% include 'main/edit_recipe.html' %} 
-    </div>
-    
-    <div id="content_to_blur">
-    <div id="filters">
-    <h3>Filtry</h3>
-    <form method="POST" id="filters_id" onsubmit="event.preventDefault();apply_filters();fetch_next(5,0)">  {% csrf_token %}
-    <h5>Nazwa przepisu</h5><div class="input_box"> {{filters.recipe_name_filter}} </div>
-    <h5>Typ herbaty</h5> <div class="input_box"> {{filters.tea_name_filter}} </div>
-    <h5>Dodatkowy składnik #1</h5>
-    <div class="input_box"> {{filters.ingredient1_filter}} </div>
-    <h5>Dodatkowy składnik #2</h5> 
-    <div class="input_box"> {{filters.ingredient2_filter}} </div>
-    <h5>Dodatkowy składnik #3</h5>
-    <div class="input_box"> {{filters.ingredient3_filter}} </div>
-    <h5>Temperatura parzenia</h5>
-    <div class="range_filter input_box">
-        <div class="range-input"> {{filters.brewing_temperatue_down_filter}}  <span class="unit unit_cel">&#8451</span> </div> 
-        <div class="separator">-</div>
-        <div class="range-input"> {{filters.brewing_temperatue_up_filter}}  <span class="unit unit_cel">&#8451</span> </div>
-    </div>
-    <h5>Czas parzenia</h5>
-    <div class="range_filter input_box">
-        <div class="range-input"> {{filters.brewing_time_down_filter}} <span class="unit">s</span> </div>
-        <div class="separator">-</div>
-        <div class="range-input"> {{filters.brewing_time_up_filter}} <span class="unit">s</span> </div>
-    </div>
-    <h5>Czas mieszania</h5>
-    <div class="range_filter input_box">
-        <div class="range-input"> {{filters.mixing_time_down_filter}} <span class="unit">s</span> </div>
-        <div class="separator">-</div>
-        <div class="range-input"> {{filters.mixing_time_up_filter}} <span class="unit">s</span> </div>
-    </div>
-    
-    <button type="submit" class="filter_btn"> Filtruj </button> <button type="reset" class="filter_btn"> Resetuj filtry</button>
-    </form>
-    </div>
-
-    <div class="arrow_box" id="left_bar">
-        <div id="left_arrow_box"><img src="{% static 'main/media/left-arrow.png' %}" id="left_arrow" class="arrow" alt="Left" onclick="fetch_next(-5,0)"> </div>
-        <div id="right_arrow_box"><img src="{% static 'main/media/right-arrow.png' %}" id="right_arrow" class="arrow" alt="Right" onclick="fetch_next(5,0)"> </div>
-    </div>
-
-
-    <div id="recipesDisplay">
-    </div>
-
-    </div>
-
-{%endblock main_content%}
-
-
-{%block additional_scripts%}
-{% csrf_token %}
-    <script>
-    const csrftoken = document.querySelector('[name=csrfmiddlewaretoken]').value;
-    var last = 0
-    </script>
-    <script src="static/main/scripts/home_scripts.js"> </script>
-    <script src="static/main/scripts/edit_recipes_scripts.js"> </script>
-    <script> fetched_recipes = {{ recipes_list|length  }} </script>
-    <script> load_options();</script>
-    <script>
-        loadMachineInfo(csrftoken);
-        fetch_next(5,0);
-    </script>
-{%endblock additional_scripts%}
