+++ src/templates/main/home.html
@@ -1,11 +1,4 @@
 {%extends "main/base.html"%}
-
-{% load static %}
-
-{% block additional_styles %}
-    <link rel="stylesheet" type="text/css" href="{% static 'main/home_styles.css' %}">
-{% endblock additional_styles %}
-
 {%block main_content%}
 <div class="mainContent" style="width: 100%; height: 100%;">
     <div class="mainTop" style="width: 100%; height: 10%;">
@@ -44,7 +37,7 @@
 </div>
 {%endblock main_content%}
 
-{%block additional_scripts%}
+{%block scripts%}
 {% csrf_token %}
 <script>
 const csrftoken = document.querySelector('[name=csrfmiddlewaretoken]').value;
@@ -53,6 +46,6 @@ var last = 0
 <script src="static/main/scripts/home_scripts.js"> </script>
 <script>
     loadRecipes(1, 3, csrftoken)
-    loadMachineInfo(csrftoken,{{ adr }})
+    loadMachineInfo(csrftoken)
 </script>
-{%endblock additional_scripts%}
+{%endblock scripts%}
