+++ src/templates/main/menu.html
@@ -1,20 +1,20 @@
 
 <div style="width:90%;height:auto;margin:20px; margin-top: 5%; margin-bottom: 5%;"></div>
 <div class="centerIn fill" style="height: 7%;">
-    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href="{% url 'app:main_home' %}"  type="button">
+    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href=''  type="button">
         Przygotuj herbatę
     </button><br />
 </div>
 
 <div class="centerIn fill" style="height: 7%;">
-    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href="{% url 'app:main_edit_tea' %}"  type="button">
+    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href=''  type="button">
         Edytuj przepisy
     </button><br />  
 </div>
 
 
 <div class="centerIn fill" style="height: 7%;">
-    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href="{% url 'app:main_browse_recipes' %}"  type="button">
+    <button class = "in" style="width: 90%; height: 90%;" onclick=location.href=''  type="button">
         Przeglądaj przepisy
     </button><br />  
 </div>
